requirements.txt
requests
beautifulsoup4
langchain
langgraph
PyPDF2
tqdm
ollama
langchain-groq
langchain-huggingface
langchain-chroma
langchain-community
python-dotenv
chromadb
bs4
lxml
pypdf
sentence-transformers
transformers
torch
pandas
fastapi
uvicorn